+++ src/Scripts/PowersLogic.lua
@@ -10,7 +10,7 @@ end)
 modutil.mod.Path.Wrap("AttachCastAtLocation", function(base, triggerArgs)
 	-- To prevent the launched Cast from the Hades boon from sticking to CrusherUnit enemies
 	if game.CurrentRun.ModsNikkelMHadesBiomesIsModdedRun and triggerArgs ~= nil and triggerArgs.TriggeredByTable ~= nil and game.GetGenusName(triggerArgs.TriggeredByTable) == "CrusherUnit" then
-		local enemyLocation = GetLocation({ Id = triggerArgs.triggeredById })
+		local enemyLocation = GetLocation({ Id = triggerArgs.triggeredById }) or {}
 		triggerArgs.LocationX = enemyLocation.X or triggerArgs.LocationX
 		triggerArgs.LocationY = enemyLocation.Y or triggerArgs.LocationY
 		triggerArgs.triggeredById = nil
@@ -18,4 +18,11 @@ modutil.mod.Path.Wrap("AttachCastAtLocation", function(base, triggerArgs)
 	end
 
 	return base(triggerArgs)
+end)
+
+modutil.mod.Path.Wrap("HadesInvisibility", function(base)
+	-- Hades reaction voicelines
+	game.thread(game.PlayVoiceLines, game.GlobalVoiceLines.ModsNikkelMHadesBiomes_HadesInvisibilityReactionVoiceLines)
+
+	return base()
 end)