+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 Other fixes and improvements:
 - You can now salute story NPCs you meet along your journey.
 - Hades and Theseus will now react with specific voicelines for certain Gods when using Godsent Hex effects against them.
+- Hades will now react to going Dark using "Unseen Ire" in the fight against him.
 - Melinoë will now react to Hades calling Cerberus to his aid.
 - Balancing: Elyisian Shades no longer create a respawning Exalted Shade if they are killed while polymorphed.
 - Balancing: Slightly reduced the chance for the "Popper" elite perk to appear.